Revised location of files adding setup-env as a stand alone miq option
